public_app/requirements.txt
streamlit
faiss-cpu
streamlit-authenticator
python-dotenv
sentence-transformers
langchain-google-genai
langchain-community
pyyaml
nest-asyncio
langchain-huggingface
